# swagger.yaml
openapi: 3.0.0
info:
  title: Mi API Aventura
  description: API educativa con autenticación, MongoDB y despliegue.
  version: 1.0.0

servers:
  - url: https://mi-api-aventura.onrender.com
    description: Producción (Render)

paths:
  /api/v1/saludo:
    get:
      summary: Saludo personalizado
      parameters:
        - name: nombre
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Saludo generado
        '400':
          description: Faltó el parámetro nombre

  /api/v1/login:
    post:
      summary: Autenticación de usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Devuelve un token JWT
        '401':
          description: Credenciales inválidas

  /api/v1/usuarios:
    get:
      summary: Obtener lista de usuarios (protegido con token)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de usuarios
        '401':
          description: Token no proporcionado
        '403':
          description: Token inválido

    post:
      summary: Crear un nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                email:
                  type: string
      responses:
        '201':
          description: Usuario creado
        '400':
          description: Datos incompletos

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT